requirements.txt
langchain-groq
langchain-core
python-dotenv
streamlit
langchain_community
langserve 
